refactor: use MockCompletionQueue from g-c-cpp-common
