+++ google/cloud/bigtable/async_list_clusters_test.cc
@@ -14,13 +14,13 @@
 
 #include "google/cloud/bigtable/admin_client.h"
 #include "google/cloud/bigtable/instance_admin.h"
-#include "google/cloud/bigtable/testing/mock_completion_queue.h"
 #include "google/cloud/bigtable/testing/mock_instance_admin_client.h"
 #include "google/cloud/bigtable/testing/mock_response_reader.h"
 #include "google/cloud/bigtable/testing/table_test_fixture.h"
 #include "google/cloud/bigtable/testing/validate_metadata.h"
 #include "google/cloud/testing_util/assert_ok.h"
 #include "google/cloud/testing_util/chrono_literals.h"
+#include "google/cloud/testing_util/mock_completion_queue.h"
 #include <gmock/gmock.h>
 #include <thread>
 
@@ -37,6 +37,7 @@ using namespace ::testing;
 using MockAsyncListClustersReader =
     google::cloud::bigtable::testing::MockAsyncResponseReader<
         btproto::ListClustersResponse>;
+using google::cloud::testing_util::MockCompletionQueue;
 
 using Functor =
     std::function<void(CompletionQueue&, ClusterList&, grpc::Status&)>;
@@ -46,7 +47,7 @@ std::string const kProjectId = "the-project";
 class AsyncListClustersTest : public ::testing::Test {
  public:
   AsyncListClustersTest()
-      : cq_impl_(new bigtable::testing::MockCompletionQueue),
+      : cq_impl_(new MockCompletionQueue),
         cq_(cq_impl_),
         client_(new testing::MockInstanceAdminClient),
         metadata_update_policy_("my_instance", MetadataParamTypes::NAME),
@@ -62,7 +63,7 @@ class AsyncListClustersTest : public ::testing::Test {
     user_future_ = instance_admin.AsyncListClusters(cq_, "my_instance");
   }
 
-  std::shared_ptr<bigtable::testing::MockCompletionQueue> cq_impl_;
+  std::shared_ptr<MockCompletionQueue> cq_impl_;
   CompletionQueue cq_;
   std::shared_ptr<testing::MockInstanceAdminClient> client_;
   future<StatusOr<ClusterList>> user_future_;
