+++ google/cloud/bigtable/internal/async_longrunning_op_test.cc
@@ -15,12 +15,12 @@
 #include "google/cloud/bigtable/internal/async_longrunning_op.h"
 #include "google/cloud/bigtable/admin_client.h"
 #include "google/cloud/bigtable/testing/mock_admin_client.h"
-#include "google/cloud/bigtable/testing/mock_completion_queue.h"
 #include "google/cloud/bigtable/testing/mock_instance_admin_client.h"
 #include "google/cloud/bigtable/testing/mock_response_reader.h"
 #include "google/cloud/bigtable/testing/table_test_fixture.h"
 #include "google/cloud/testing_util/assert_ok.h"
 #include "google/cloud/testing_util/chrono_literals.h"
+#include "google/cloud/testing_util/mock_completion_queue.h"
 #include <google/bigtable/v2/bigtable.pb.h>
 #include <gmock/gmock.h>
 #include <thread>
@@ -39,6 +39,7 @@ using google::bigtable::v2::SampleRowKeysResponse;
 using MockAsyncLongrunningOpReader =
     google::cloud::bigtable::testing::MockAsyncResponseReader<
         google::longrunning::Operation>;
+using google::cloud::testing_util::MockCompletionQueue;
 
 void OperationFinishedSuccessfully(google::longrunning::Operation& response,
                                    grpc::Status& status) {
@@ -58,7 +59,7 @@ class AsyncLongrunningOpFutureTest : public bigtable::testing::TableTestFixture,
 TEST_P(AsyncLongrunningOpFutureTest, EndToEnd) {
   auto const success = GetParam();
   auto client = std::make_shared<testing::MockAdminClient>();
-  auto cq_impl = std::make_shared<testing::MockCompletionQueue>();
+  auto cq_impl = std::make_shared<MockCompletionQueue>();
   bigtable::CompletionQueue cq(cq_impl);
 
   auto longrunning_reader =
@@ -141,7 +142,7 @@ class AsyncLongrunningOperationTest : public ::testing::Test {
  public:
   AsyncLongrunningOperationTest()
       : client_(std::make_shared<testing::MockAdminClient>()),
-        cq_impl_(std::make_shared<testing::MockCompletionQueue>()),
+        cq_impl_(std::make_shared<MockCompletionQueue>()),
         cq_(cq_impl_),
         longrunning_reader_(google::cloud::internal::make_unique<
                             MockAsyncLongrunningOpReader>()),
@@ -192,7 +193,7 @@ class AsyncLongrunningOperationTest : public ::testing::Test {
   }
 
   std::shared_ptr<testing::MockAdminClient> client_;
-  std::shared_ptr<testing::MockCompletionQueue> cq_impl_;
+  std::shared_ptr<MockCompletionQueue> cq_impl_;
   bigtable::CompletionQueue cq_;
   std::unique_ptr<MockAsyncLongrunningOpReader> longrunning_reader_;
   std::unique_ptr<grpc::ClientContext> context_;
