+++ google/cloud/bigtable/internal/async_retry_multi_page_test.cc
@@ -15,13 +15,13 @@
 #include "google/cloud/bigtable/internal/async_retry_multi_page.h"
 #include "google/cloud/bigtable/instance_admin_client.h"
 #include "google/cloud/bigtable/table.h"
-#include "google/cloud/bigtable/testing/mock_completion_queue.h"
 #include "google/cloud/bigtable/testing/mock_instance_admin_client.h"
 #include "google/cloud/bigtable/testing/mock_response_reader.h"
 #include "google/cloud/bigtable/testing/mock_sample_row_keys_reader.h"
 #include "google/cloud/bigtable/testing/table_test_fixture.h"
 #include "google/cloud/testing_util/assert_ok.h"
 #include "google/cloud/testing_util/chrono_literals.h"
+#include "google/cloud/testing_util/mock_completion_queue.h"
 #include <google/bigtable/admin/v2/bigtable_instance_admin.grpc.pb.h>
 #include <future>
 #include <thread>
@@ -37,6 +37,7 @@ namespace bt = ::google::cloud::bigtable;
 namespace btproto = google::bigtable::v2;
 using namespace ::testing;
 using namespace google::cloud::testing_util::chrono_literals;
+using google::cloud::testing_util::MockCompletionQueue;
 
 class BackoffPolicyMock : public bigtable::RPCBackoffPolicy {
  public:
@@ -102,7 +103,7 @@ class AsyncMultipageFutureTest : public ::testing::Test {
             bigtable::DefaultRPCRetryPolicy(internal::kBigtableLimits)),
         shared_backoff_policy_mock_(
             google::cloud::internal::make_unique<SharedBackoffPolicyMock>()),
-        cq_impl_(new bigtable::testing::MockCompletionQueue),
+        cq_impl_(new google::cloud::testing_util::MockCompletionQueue),
         cq_(cq_impl_),
         client_(new testing::MockInstanceAdminClient),
         metadata_update_policy_("my_instance", MetadataParamTypes::NAME) {}
@@ -187,7 +188,7 @@ class AsyncMultipageFutureTest : public ::testing::Test {
  protected:
   std::unique_ptr<RPCRetryPolicy> rpc_retry_policy_;
   std::unique_ptr<SharedBackoffPolicyMock> shared_backoff_policy_mock_;
-  std::shared_ptr<bigtable::testing::MockCompletionQueue> cq_impl_;
+  std::shared_ptr<google::cloud::testing_util::MockCompletionQueue> cq_impl_;
   CompletionQueue cq_;
   std::shared_ptr<testing::MockInstanceAdminClient> client_;
   MetadataUpdatePolicy metadata_update_policy_;
